revert type change for rpt_period_id changes

+++ src/entities/workspace/emission-evaluation.entity.ts
@@ -20,7 +20,7 @@ export class EmissionEvaluation extends BaseEntity {
   @PrimaryColumn({
     name: 'rpt_period_id',
     transformer: new NumericColumnTransformer(),
-    type: 'bigint',
+    type: 'numeric',
     nullable: false,
   })
   reportingPeriodId: number;
