+++ src/entities/workspace/long-term-fuel-flow.entity.ts
@@ -20,7 +20,7 @@ export class LongTermFuelFlow extends BaseEntity {
   @Column({
     name: 'rpt_period_id',
     transformer: new NumericColumnTransformer(),
-    type: 'bigint',
+    type: 'numeric',
     nullable: true,
   })
   reportingPeriodId: number;
